minor: code refactoring to isolate functionality
