+++ components/PotreeContainer.vue
@@ -8,7 +8,7 @@
         <div class="btn " @click="toggleSidebar">
           Toggle Panel
         </div>
-        <div class="btn" @click="moveCamera">
+        <div class="btn" @click="animateCamera(0)">
           move camera
         </div>
         <div class="btn" @click="toggleAnimationVisibility">
@@ -16,17 +16,19 @@
         </div>
       </div>
     </div>
-    <camera-section v-if="camera" :active-camera="camera" :position="position" :view="view" />
+    <camera-section v-if="camera" />
   </div>
 </template>
 
 <script>
 // import Vue from 'vue'
-import { onMounted, ref } from '@nuxtjs/composition-api'
+import { ref } from '@nuxtjs/composition-api'
 import { OBJLoader } from 'three/examples/jsm/loaders/OBJLoader.js'
 import { VAOrientedImageLoader } from './overrides/VAOrientedImages'
-import { VAFirstPersonControls } from './overrides/VAFirstPersonControls'
-import { potreeRef } from '~/scene/VAPotree'
+import {
+  animateCamera, potreeRef, toggleAnimationVisibility, addAnimationPath,
+  setInitialSceneParameters, loadInitialPointCloud
+} from '~/API/VAPotree'
 
 // Access the potreeView instance from everywhere using composition API
 // export const potreeRef = reactive({})
@@ -34,192 +36,70 @@ export const isSidebarOpen = ref(false)
 
 export default {
   name: 'PotreeViewer',
-  props: {
-    graphics: {
-      type: String,
-      required: false,
-      default: 'medium',
-      validator (value) {
-        return ['low', 'medium', 'high'].includes(value)
-      }
-    },
-    numPoints: {
-      type: Number,
-      required: false,
-      default: 6000000,
-      validator (value) {
-        return value > 0 && value < 50000000
-      }
-    },
-    pointClouds: {
-      type: Array,
-      required: false,
-      default: () => []
-    }
-  },
 
   setup (props, context) {
-    // const containerRef = ref(null)
-    // console.log('setup', props)
-    // const potree = ref(null)
-    onMounted(() => {
-      // check if the sidebar is visible
-      isSidebarOpen.value = $('#potree_sidebar_container').is(':visible')
-
-      // Logs: `Headline`
-
-      // console.log('🎹', containerRef)
-    })
-
-    /* Tween function used with camera movement - arrow buttons
-*****************************************************/
-    // function tweenFunction (direction, currentAmountObject, newAmountObject) {
-    //   // Setup the animation loop.
-    //   function animate (time) {
-    //     requestAnimationFrame(animate)
-    //     TWEEN.update(time)
-    //   }
-    //   requestAnimationFrame(animate)
-    //
-    //   // declare and initalize tween
-    //   const tween = new TWEEN.Tween(currentAmountObject)
-    //     .to(newAmountObject, 1000)
-    //     .easing(TWEEN.Easing.Quintic.Out)
-    //     .onUpdate(function () {
-    //       if (direction === 'up' || direction === 'down') {
-    //         potreeRef.viewer.scene.view.pitch = currentAmountObject.amount
-    //       } else if (direction === 'left' || direction === 'right') {
-    //         potreeRef.viewer.scene.view.yaw = currentAmountObject.amount
-    //       }
-    //     })
-    //     .start() // Start the tween immediately.
-    // }
-
-    /* Rotate Camera Up
-    *****************************************************/
-    // function rotateUp () {
-    //   // movement direction
-    //   const direction = 'up'
-    //
-    //   // NOTE: we need values as a property in an object for tween to work
-    //   const currentPitchObject = {
-    //     amount: potreeRef.viewer.scene.view.pitch
-    //   }
-    //   const newPitchObject = {
-    //     amount: potreeRef.viewer.scene.view.pitch + 0.25
-    //   }
-    //
-    //   // call tween function
-    //   tweenFunction(direction, currentPitchObject, newPitchObject)
-    // }
-
-    // Methods
-    function moveCamera ({ x = 296264.07, y = 4633679, z = 129 }) {
-      console.log('🎹', x, y, z, potreeRef.viewer.scene)
-      console.log('🎹', potreeRef.viewer.scene.view.getPivot().toArray())
-
-      // fly to: the animation
-      potreeRef.viewer.scene.cameraAnimations[0].play()
-    }
-    function toggleAnimationVisibility () {
-      potreeRef.viewer.scene.view.getPivot().toArray()
-      // TODO this only does it for the first element in the array
-      potreeRef.viewer.scene.cameraAnimations[0].setVisible(!potreeRef.viewer.scene.cameraAnimations[0].visible)
-      console.log('🎹', potreeRef.viewer.scene.cameraAnimations[0])
-
-      // fly to:
-    }
-
-    // animation paths
-    function addAnimationPath () {
-      const animation = new Potree.CameraAnimation(potreeRef.viewer)
-
-      const positions = [
-        [296267.12792342174, 4633681.40622494, 128.3653329922611],
-        [296271.14629413467, 4633685.636950317, 129.3653329922611],
-        [296286.3704886272, 4633683.871486212, 132.01175511086956]
-      ]
-
-      const targets = [
-        [296250.7719238924, 4633714.095726619, 126.43484628061825],
-        [296235.8339610568, 4633681.722820112, 125.17872130141684],
-        [296235.8339610568, 4633681.732820112, 125.17872130141684]
-      ]
-
-      for (let i = 0; i < positions.length; i++) {
-        const cp = animation.createControlPoint()
-
-        cp.position.set(...positions[i])
-        cp.target.set(...targets[i])
-      }
-
-      potreeRef.viewer.scene.addCameraAnimation(animation)
-      // animation.play()
-    }
-
     return {
-      // It is important to return the ref,
-      // otherwise it won't work.
-      // potreeRef,
       toggleAnimationVisibility,
-      addAnimationPath,
       isSidebarOpen,
-      moveCamera
+      animateCamera
     }
   },
 
   data () {
     return {
-      position: { x: 0, y: 0, z: 0 },
-      camera: null,
-      view: null,
-      offset: 0,
-      activeImage: null
+      camera: null
     }
   },
   mounted () {
+    // check if the sidebar is visible
+    isSidebarOpen.value = $('#potree_sidebar_container').is(':visible')
+
     // hide toolbar if production mode
     if (!this.$nuxt.context.isDev) {
       this.toggleSidebar()
     }
-    // const s = new Potree.Viewer(this.$refs.potree_container)
-    // cosa.viewer = new Potree.Viewer(this.$refs.potree_container)
 
-    // Vue.prototype.$viewer = new Potree.Viewer(this.$refs.potree_container)
-    potreeRef.viewer = new Potree.Viewer(this.$refs.potree_container) // this not, adn I think it is because it is a recursive error object?
-    // we need to pass to the global value the viewer, otherwise, the animation won't be able to load
-    window.viewer = potreeRef.viewer
-    // const viewer = new Potree.Viewer(this.$refs.potree_container)      //// THIS WORKS
+    // Load potree viewer inside the DOM
+    potreeRef.viewer = new Potree.Viewer(this.$refs.potree_container)
+    setInitialSceneParameters()
+    loadInitialPointCloud(this.$nuxt.context.isDev)
 
     const viewer = potreeRef.viewer
-    // CLEAN UP THE MESS WITH THE VARIABLES!
     const scene = viewer.scene
+    this.camera = scene.getActiveCamera()
 
-    scene.annotations.add(
-      new Potree.Annotation({
-        position: [296264.396, 4633679.776, 129.778],
-        title: 'About Annotations',
+    // Example Add image to the scene
+    {
+      const cameraParamsPath = '/images/images.xml'
+      const imageParamsPath = '/images/pyramid.txt'
+      VAOrientedImageLoader.load(cameraParamsPath, imageParamsPath, viewer)
+        .then(([images, controls]) => {
+          viewer.scene.addOrientedImages(images)
+          const material = this.createMaterial()
+          material.transparent = false
+          images.images[0].mesh.material = material
+        })
+    }
 
-        cameraPosition: [296264.39688606694, 4633679.776566018, 129.77835768357866],
-        cameraTarget: [296254.64710414334, 4633692.914354076, 127.63586441697944],
-        description: `
-<ul>
-  <li>Click on the annotation label to move a predefined view.</li>
-  <li>Click on the icon to execute the specified action.
-  </li>In this case, the action will bring you to another scene and point cloud.</ul>
-`
-      }))
-
-    this.addAnimationPath(viewer)
+    // Example annotation
+    {
+      const annotationName = new Potree.Annotation({
+        position: [296249.19705492083, 4633726.448203472, 128.7690558114301],
+        title: 'Piramide',
+        cameraPosition: [296260.6092322839, 4633696.311992192, 130.7733116269969],
+        cameraTarget: [296255.42565172265, 4633711.747814068, 133.44087523067427],
+        description: 'Click on the annotation label to move a predefined view. <br>Click on the icon to execute the specified action.<br>In this case, the action will bring you to another scene and point cloud.'
+      })
+      scene.annotations.add(annotationName)
+      addAnimationPath()
+    }
 
     // LIGHTS
     {
       const directional = new THREE.DirectionalLight(0xFFFFFF, 1.0)
       directional.position.set(10, 10, 10)
       directional.lookAt(0, 0, 0)
-
       const ambient = new THREE.AmbientLight(0x555555)
-
       viewer.scene.scene.add(directional)
       viewer.scene.scene.add(ambient)
     }
@@ -273,160 +153,13 @@ export default {
         })
       }, onProgress, onError)
     }
-
-    // Get active camera position
-    this.camera = scene.getActiveCamera()
-    this.view = scene.view
-
-    // Set the position
-    this.position = this.camera.position
-
-    viewer.setFOV(60)
-    viewer.setBackground('skybox')
-    viewer.setEDLEnabled(false)
-    viewer.setPointBudget(3_000_000)
-    viewer.loadSettingsFromURL()
-
-    // Pointcloud data source
-    const POINT_CLOUD_URL = this.$nuxt.context.isDev
-      // locally
-      ? 'http://localhost:3000/pointclouds/DRIVE_1_V3_levels_8/cloud.js'
-      // Cloud storage
-      : 'https://storage.googleapis.com/via-appia-20540.appspot.com/cloud.js'
-
-    // hide menu button in the sidebar
-    $('#potree_quick_buttons').hide()
-
-    Potree.loadPointCloud(
-      POINT_CLOUD_URL,
-      'Drive Map',
-      ({ pointcloud }) => {
-        pointcloud.material.size = 1
-        pointcloud.material.pointSizeType = Potree.PointSizeType.ADAPTIVE
-        pointcloud.material.shape = Potree.PointShape.SQUARE
-
-        // Load pointcloud data
-        viewer.scene.addPointCloud(pointcloud)
-
-        // Set initial camera view position
-        viewer.scene.view.position.set(
-          296264.39688606694,
-          4633679.776566018,
-          129.77835768357866
-        )
-        viewer.scene.view.yaw = 0.3
-        viewer.scene.view.pitch = 0
-
-        // Control camera with the keyboard
-        viewer.fpControls = new VAFirstPersonControls(viewer)
-        viewer.fpControls.addEventListener('start', viewer.disableAnnotations.bind(viewer))
-        viewer.fpControls.addEventListener('end', viewer.enableAnnotations.bind(viewer))
-        viewer.setControls(viewer.fpControls)
-        viewer.setMoveSpeed(5.5)
-
-        const cameraParamsPath = '/images/images.xml'
-        const imageParamsPath = '/images/pyramid.txt'
-
-        VAOrientedImageLoader.load(
-          cameraParamsPath,
-          imageParamsPath,
-          viewer
-        ).then(([images, controls]) => {
-          viewer.scene.addOrientedImages(images)
-
-          // const material = this.createMaterial()
-          // material.transparent = true
-          // images.images[0].mesh.material = material
-        })
-      }
-    )
-
-    viewer.loadGUI(() => {
-      viewer.setLanguage('en')
-      $('#menu_tools').next().show()
-      $('#menu_clipping').next().show()
-      $('#menu_scene').next().show()
-
-      // Add custom section for Camera
-      const cameraSection = $(`
-        <h3 id="menu_camera" class="accordion-header ui-widget"><span>Camera</span></h3>
-        <div class="accordion-content ui-widget pv-menu-list"></div>
-        `)
-
-      // get vue component for Camera Section
-      const cameraSectionHTML = document.getElementById('cameraSection')
-      const cameraSectionContent = cameraSection.last()
-      cameraSectionContent.html(cameraSectionHTML)
-      cameraSection.first().click(() => cameraSectionContent.slideToggle())
-      cameraSection.insertBefore($('#menu_tools'))
-
-      viewer.toggleSidebar()
-    })
-    viewer.addEventListener('move_speed_changed', () => {
-      // Set the position
-      this.position = this.camera.position
-    })
   },
   methods: {
     toggleSidebar () {
       $('#potree_sidebar_container').toggle()
       isSidebarOpen.value = $('#potree_sidebar_container').is(':visible')
-    },
-
-    // createMaterial() {
-    //   let vertexShader = `
-    //     uniform float uNear;
-    //     varying vec2 vUV;
-    //     varying vec4 vDebug;
-
-    //     void main(){
-    //       vDebug = vec4(0.0, 1.0, 0.0, 1.0);
-    //       vec4 modelViewPosition = modelViewMatrix * vec4(position, 1.0);
-    //       // make sure that this mesh is at least in front of the near plane
-    //       modelViewPosition.xyz += normalize(modelViewPosition.xyz) * uNear;
-    //       gl_Position = projectionMatrix * modelViewPosition;
-    //       vUV = uv;
-    //     }`;
-
-    //   let fragmentShader = `
-    //     uniform sampler2D tColor;
-    //     uniform float uOpacity;
-    //     varying vec2 vUV;
-    //     varying vec4 vDebug;
-    //     void main(){
-    //       vec4 color = texture2D(tColor, vUV);
-    //       gl_FragColor = color;
-    //       gl_FragColor.a = uOpacity;
-    //     }`;
-
-    //   const material = new THREE.ShaderMaterial({
-    //     uniforms: {
-    //       // time: { value: 1.0 },
-    //       // resolution: { value: new THREE.Vector2() }
-    //       tColor: { value: new THREE.Texture() },
-    //       uNear: { value: 0.0 },
-    //       uOpacity: { value: 0.5 }
-    //     },
-    //     vertexShader: vertexShader,
-    //     fragmentShader: fragmentShader,
-    //     side: THREE.DoubleSide
-    //   });
-
-    //   material.side = THREE.DoubleSide;
-
-    //   return material;
-    // },
-
-    createAnnotations () {
-      this.$viewer.scene.annotations.children = []
-      this.$viewer.scene.addAnnotation([236790, 548513, 69], {
-        // title: this.$t('commanderHouse')
-        title: 'commanderHouse'
-      })
-      this.$viewer.scene.addAnnotation([237079, 548442, 69], {
-        title: 'campTerrain'
-      })
     }
+
   }
 }
 </script>
