--- scene/VAPotree.js
@@ -1,4 +1,0 @@
-// Access the potreeView instance from everywhere using composition API
-import { reactive } from '@nuxtjs/composition-api'
-
-export const potreeRef = reactive({})
